virtcontainers: Split the factory package into Linux and Darwin bits
- split template
- split factory
- add stubs for darwin
