+++ src/runtime/virtcontainers/factory/template/template_darwin.go
@@ -1,0 +1,10 @@
+//
+// Copyright (c) 2022 Apple, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// template implements base vm factory with vm templating.
+
+package template
+
+type template struct{}
